% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MSGFgui}
\alias{MSGFgui}
\title{Start MSGFgui in the default browser}
\usage{
MSGFgui()
}
\value{
This function interupts the R session until the user press escape and
has no return value.
}
\description{
This function load the GUI in the users default browser.
}

