% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MSGFgui.R
\name{currentData}
\alias{currentData}
\title{Gets the result data currently in use by another R session running MSGFgui}
\usage{
currentData()
}
\value{
An mzIDCollection object containing the current identification
results in the the MSGFgui browser
}
\description{
This function makes it possible to directly access the data used in the gui
by starting a new R session and running this function. This will only work if
just one instance of MSGFgui is running in the browser. The data will linger
on so calling \code{currentData()} when no instance of MSGFgui is running
will return the last set of data used.
}
\examples{
results <- currentData()
}

