% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MSGFgui}
\alias{MSGFgui}
\title{Start MSGFgui in the default browser}
\usage{
MSGFgui(...)
}
\arguments{
\item{...}{Parameters passed onto \code{\link[shiny]{runApp}}}
}
\value{
This function interupts the R session until the user press escape and
has no return value.
}
\description{
This function load the GUI in the users default browser. Due to some
unfortunate problems with shiny, the user needs to refresh the webpage the
first time it loads in order for everything to work properly. This will
likely be fixed in a coming shiny release. For the time being MSGFgui is
designed to run locally. This means that there is automatic view rights to
the full file system of the computer running the server.
}
\examples{
\donttest{
# Default mode
MSGFgui()

# Pass on parameters to shiny
MSGFgui(host='0.0.0.0')
}
}

